sending json and arranged product server and client
