+++ client/scripts/cancel.js
@@ -1,3 +1,4 @@
 function cancel() {   // Function for cancel button in Add Products
-    location.reload();
+  document.getElementById("form").className = "show";
+  document.getElementById("formAdd").className = "hidden";
   }
