+++ client/views/form.html
@@ -1,0 +1,9 @@
+
+<script>
+// function to loading page with the product registered in DB
+showProduct(); 
+</script>
+
+
+          
+
