--- client/views/form.php
@@ -1,6 +1,0 @@
-<div id="container" class="container"> 
-
-    <?PHP require ("server/api/product.php"); ?>
-
-</div>
-
